+++ packages/hightide-utils/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.7] - 2026-09-18
+
+### Added
+
+- Translation keys `allowedFileTypes` and `maximumNumberOfFiles`
+
 ## [0.1.6] - 2026-09-16
 
 ### Added